# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Taeyoung",
        "description": "API Documentation for Taeyoung",
        "version": "1.0.0"
    },
    "paths": {
...
        "/taeyoung-backend/public/api/bank": {
            "get": {
                "tags": [
                    "Bank"
                ],
                "summary": "Get all banks with pagination",
                "operationId": "976da2d0ca11a43e5fe9bd66946ae3f9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by column",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank collection response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BankCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bank"
                ],
                "summary": "Store a newly created bank in storage.",
                "operationId": "f57115dde8f8fbdbe296ada2f76498f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BankRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bank created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bank"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The name has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/taeyoung-backend/public/api/bank/{id}": {
            "get": {
                "tags": [
                    "Bank"
                ],
                "summary": "Display the specified bank.",
                "operationId": "e78b4d73a80a7d06ba364fa6839064ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bank"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bank not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bank"
                ],
                "summary": "Update the specified bank in storage.",
                "operationId": "98131e8eaf955e0c802e8ad689a2ebef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BankRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bank updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bank"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bank not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The name has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bank"
                ],
                "summary": "Remove the specified bank from storage.",
                "operationId": "ec3fa796501ed4471eb0fabfe58df288",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bank deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bank not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Bank related to payment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bank related to payment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/taeyoung-backend/public/api/path?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/taeyoung-backend/public/api/path?page=4"
                    },
                    "prev": {
                        "type": "string",
                        "example": "null"
                    },
                    "next": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/taeyoung-backend/public/api/path?page=2"
                    }
                },
                "type": "object"
            },
            "PaginationMetaLinks": {
                "properties": {
                    "url": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/taeyoung-backend/public/api/path?page=1"
                    },
                    "label": {
                        "type": "string",
                        "example": "1"
                    },
                    "active": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": "1"
                    },
                    "from": {
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "example": "4"
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationMetaLinks"
                    },
                    "path": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/taeyoung-backend/public/api/path"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": "15"
                    },
                    "to": {
                        "type": "integer",
                        "example": "15"
                    },
                    "total": {
                        "type": "integer",
                        "example": "60"
                    }
                },
                "type": "object"
            },
...
            "BankResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Banco de la Nación"
                    }
                },
                "type": "object"
            },
            "BankCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BankResource"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
...
            "Bank": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Bank BCA"
                    }
                },
                "type": "object"
            },
            "BankRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Bank BCA"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}